+++ Agents/codeGeneration/helperFunctions.py
@@ -1,0 +1,48 @@
+from langchain_google_genai import ChatGoogleGenerativeAI
+from langchain.agents import AgentExecutor, create_tool_calling_agent
+from langchain_core.prompts import ChatPromptTemplate, PromptTemplate, SystemMessagePromptTemplate
+
+def create_tool_agent(llm: ChatGoogleGenerativeAI, tools: list, system_prompt: str):
+    """Helper function to create agents with custom tools and system prompt
+    Args:
+        llm (ChatVertexAI): LLM for the agent
+        tools (list): list of tools the agent will use
+        system_prompt (str): text describing specific agent purpose
+
+    Returns:
+        executor (AgentExecutor): Runnable for the agent created.
+    """
+    
+    # Each worker node will be given a name and some tools.
+    
+    system_prompt_template = PromptTemplate(
+
+                template= system_prompt + """
+                ONLY respond to the part of query relevant to your purpose.
+                IGNORE tasks you can't complete. 
+                """
+            )
+
+    #define system message
+    system_message_prompt = SystemMessagePromptTemplate(prompt=system_prompt_template)
+
+    human_message = PromptTemplate(
+
+                template= system_prompt + """
+                Here is the data report, 
+                based on it call the visualizations needed by following the system instruction:
+                \n\n {data_report}
+                """,
+                input_variables=["data_report"],
+            )
+    prompt = ChatPromptTemplate.from_messages(
+        [
+        system_message_prompt,
+        ('human', human_message),
+        ]
+    )
+
+    agent = create_tool_calling_agent(llm, tools, prompt)
+    executor = AgentExecutor(agent=agent, tools=tools, 
+                return_intermediate_steps= True, verbose = False)
+    return executor
