+++ Agents/codeGeneration/planner.py
@@ -1,0 +1,28 @@
+from enum import Enum
+from pydantic import BaseModel
+members = ["CodeGen", 'Visualizer']
+
+#create options map for the supervisor output parser.
+member_options = {member:member for member in members}
+
+#create Enum object
+MemberEnum = Enum('MemberEnum', member_options)
+
+
+#force Supervisor to pick from options defined above
+# return a dictionary specifying the next agent to call 
+#under key next.
+class SupervisorOutput(BaseModel):
+    #defaults to communication agent
+    next: MemberEnum = MemberEnum.Communicate
+    
+system_prompt = (
+    """You are a supervisor tasked with managing a conversation between the
+    crew of workers:  {members}. Given the following user request, 
+    and crew responses respond with the worker to act next.
+    Each worker will perform a task and respond with their results and status. 
+    When finished with the task, route to communicate to deliver the result to 
+    user. Given the conversation and crew history below, who should act next?
+    Select one of: {options} 
+    \n{format_instructions}\n"""
+)
