+++ requirements.txt
@@ -1,3 +1,6 @@
+aiohappyeyeballs==2.4.4
+aiohttp==3.11.11
+aiosignal==1.3.2
 altair==5.5.0
 annotated-types==0.7.0
 anyio==4.7.0
@@ -14,27 +17,43 @@ contourpy==1.3.1
 cycler==0.12.1
 fastapi==0.115.6
 fonttools==4.55.3
+frozenlist==1.5.0
 gevent==24.11.1
 gitdb==4.0.11
 GitPython==3.1.43
 greenlet==3.1.1
 h11==0.14.0
+httpcore==1.0.7
 httptools==0.6.4
+httpx==0.28.1
 idna==3.10
 Jinja2==3.1.4
+jsonpatch==1.33
+jsonpointer==3.0.0
 jsonschema==4.23.0
 jsonschema-specifications==2024.10.1
 kiwisolver==1.4.7
+langchain==0.3.13
+langchain-core==0.3.28
+langchain-text-splitters==0.3.4
+langgraph==0.2.60
+langgraph-checkpoint==2.0.9
+langgraph-sdk==0.1.48
+langsmith==0.2.7
 markdown-it-py==3.0.0
 MarkupSafe==3.0.2
 matplotlib==3.10.0
 mdurl==0.1.2
+msgpack==1.1.0
+multidict==6.1.0
 narwhals==1.18.4
-numpy==2.2.0
+numpy==1.26.4
+orjson==3.10.13
 packaging==24.2
 pandas==2.2.3
 pillow==11.0.0
 plotly==5.24.1
+propcache==0.2.1
 protobuf==5.29.2
 pyarrow==18.1.0
 pycparser==2.22
@@ -50,12 +69,14 @@ pytz==2024.2
 PyYAML==6.0.2
 referencing==0.35.1
 requests==2.32.3
+requests-toolbelt==1.0.0
 rich==13.9.4
 rpds-py==0.22.3
 seaborn==0.13.2
 six==1.17.0
 smmap==5.0.1
 sniffio==1.3.1
+SQLAlchemy==2.0.36
 st-theme==1.2.3
 starlette==0.41.3
 streamlit==1.41.1
@@ -74,5 +95,6 @@ watchfiles==1.0.3
 websocket==0.2.1
 websocket-client==1.8.0
 websockets==14.1
+yarl==1.18.3
 zope.event==5.0
 zope.interface==7.2
